% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiverse_status.R
\name{runiverse_ok}
\alias{runiverse_ok}
\title{Check if package builds/checks are successful in R-Universe}
\usage{
runiverse_ok(pkgName)
}
\arguments{
\item{pkgName}{\code{character(1)} The name of the package to query in the
R-Universe.}
}
\value{
\code{logical(1)} \code{TRUE} if at least one build/check status is not
\code{"ERROR"}, \code{"FAIL"}, or \code{"CANCELLED"}. Returns \code{FALSE} otherwise.
}
\description{
This function checks the status of a package in the R-Universe.
It returns \code{TRUE} if at least one build/check has a non-failure status
(i.e., not \code{"ERROR"}, \code{"FAIL"}, or \code{"CANCELLED"}).
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
runiverse_ok("BiocCheck")
\dontshow{\}) # examplesIf}
}
