% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{remove_criteria}
\alias{remove_criteria}
\title{Remove gate criteria the package is exempt from}
\usage{
remove_criteria(package, branch, criteria, manifest = .BIOCONDUCTOR_MANIFEST)
}
\arguments{
\item{package}{name}

\item{branch}{Bioconductor git branch}

\item{criteria}{list()}

\item{manifest}{(defaults to .BIOCONDUCTOR_MANIFEST) path to repository}
}
\value{
list of remaining criteria
}
\description{
Remove gate criteria the package is exempt from
}
\details{
Expects exemptions in the Bioconductor manifest repository
in a DCF-formated exemptions.txt file. A package with exemptions has
Package and Exemptions entries. The Exemptions field is a comma-separated
list of gate names, such as in \code{source_criteria()}.
}
\examples{
\dontrun{
remove_criteria("BiocCheck","devel", default_criteria())
}

}
