% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiverse_status.R
\name{runiverse_status}
\alias{runiverse_status}
\title{Check the status of a package in the R-Universe}
\usage{
runiverse_status(pkgName)
}
\description{
Check the status of a package in the R-Universe
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
runiverse_status("BiocCheck")
\dontshow{\}) # examplesIf}
}
