% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{unregister_gates}
\alias{unregister_gates}
\title{Unregister gate criteria}
\usage{
unregister_gates(criteria, gates)
}
\arguments{
\item{criteria}{A criteria list, e.g. from \code{\link[=default_criteria]{default_criteria()}}.}

\item{gates}{\code{character()} Names for removal in criteria.}
}
\value{
The updated criteria list.
}
\description{
Remove a named gate criterion in a criteria list,
}
\examples{
criteria <- default_criteria()
criteria <- unregister_gates(
    criteria, "no_large_files"
)
names(criteria$gates)

}
